apps/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]

